+++ lib/shrine/plugins/entity.rb
@@ -112,7 +112,13 @@ class Shrine
         #     attacher.file #=> #<Shrine::UploadedFile>
         def reload
           read
+          @previous = nil
           self
+        end
+
+        # Loads attachment from the entity attribute.
+        def read
+          load_column(read_attribute)
         end
 
         # Returns a hash with entity attribute name and column data.
@@ -133,11 +139,6 @@ class Shrine
         end
 
         private
-
-        # Loads attachment from the entity attribute.
-        def read
-          load_column(read_attribute)
-        end
 
         # Reads value from the entity attribute.
         def read_attribute