+++ test/plugin/sequel_test.rb
@@ -195,6 +195,28 @@ describe Shrine::Plugins::Sequel do
         assert_equal file, @user.avatar
       end
 
+      it "preserves state other than the file" do
+        @user.class.include @shrine::Attachment.new(:avatar)
+
+        @user.save
+        @user.avatar_attacher(cache: :other_cache)
+        @user.avatar_attacher.context[:foo] = "bar"
+        @user.reload
+
+        assert_equal :other_cache, @user.avatar_attacher.cache_key
+        assert_equal "bar",        @user.avatar_attacher.context[:foo]
+      end
+
+      it "doesn't initialize the attacher if it hasn't been initialized" do
+        @user.class.include @shrine::Attachment.new(:avatar)
+
+        @user.save
+        @user = @user.class.with_pk!(@user.id)
+        @user.reload
+
+        refute @user.instance_variable_defined?(:@avatar_attacher)
+      end
+
       it "returns self" do
         @user.class.include @shrine::Attachment.new(:avatar)