+++ test/test_helper.rb
@@ -1,3 +1,5 @@
+Warning[:deprecated] = true if RUBY_VERSION >= "2.7"
+
 require "bundler/setup"
 
 if ENV["COVERAGE"]